+++ src/nhp/capacity/functional_areas/classifications.py
@@ -8,3 +8,30 @@ def class_age_adult() -> Column:
 
 def class_age_child() -> Column:
     return F.col("age") < 18
+
+
+def class_ae() -> Column:
+    return F.col("aedepttype") == "01"
+
+
+def class_ae_resus() -> Column:
+    return F.col("acuity") == "immediate-resuscitation"
+
+
+def class_ae_major() -> Column:
+    return F.col("acuity") == "very-urgent"
+
+
+def class_ae_minor() -> Column:
+    return (
+        F.col("acuity").isin(
+            "standard",
+            "non-urgent",
+            "urgent",
+        )
+        | F.col("acuity").isNull()
+    )
+
+
+def class_sdec() -> Column:
+    return F.col("aedepttype") == "05"
